Refactor exports to export the things that should be exported
